+++ resources/views/coursesList.blade.php
@@ -37,7 +37,7 @@
                                href="/courses/{{$course->id}}/edit">Редактировать</a>
                         </td>
                         <td>
-                            <a class="btn btn-danger" href="/courses/{{$course->id}}/destroy">Удалить</a>
+                            <a class="confirm_delete btn btn-danger" href="/courses/{{$course->id}}/destroy">Удалить</a>
                         </td>
                     </tr>
                 @endforeach
@@ -45,4 +45,5 @@
             {{ $coursesList->links('vendor.pagination.bootstrap-4') }}
         @endif
     </div>
+    <script src="/assets/js/delete-confirm.js"></script>
 @endsection
